--- FastExp.hpp
@@ -1,5 +1,0 @@
-/* max. rel. error <= 1.73e-3 on [-87,88] */
-
-#include <cmath> 
-
-float fastexp(float x);
