+++ src/FastExp.hpp
@@ -1,0 +1,5 @@
+/* max. rel. error <= 1.73e-3 on [-87,88] */
+
+#include <cmath> 
+
+float fastexp(float x);
